Fix tests for GameMaster::evaluate_guess
The unit tests for GameMaster::evaluate_guess caught a bug with the
value of color_only. It seems like arrays weren't being handled
properly and so color_count had completely incorrect information.

To resolve the issue, we use std::vector. This seems to be cleaner
than using arrays, anyways, since we no longer need to store the
number of spaces or the number of colors.